Add feature for picking points under cursor
Followed the model from this post
https://threejsfundamentals.org/threejs/lessons/threejs-picking.html
to create functionality for picking points under the mouse.

At the moment, the id of the point under the mouse is simply
shown in the span with id = tooltip on the top of the page.
Also, the color of the whole batch containing the selected
point is changed to red. The plan is for this to evolve into
a local tooltip over the point, only highlighting the
selected point, and posting a message to the NGCHM to highlight
the point on the NGCHM, etc. etc.