Add option for 'box' or 'origin' coordinates
Most people prefer the original box-style coordinate rendering.
However some prefer the coordinates at the origin in a more
tranditional Cartesian rendering. This commit adds the option in the
gear menu for the user to choose their preference.

Also adds an option to choose the point size. (this will eventually
become a slider)