Fix rendering bug: previous choices showup
I *think* maybe this fixes a bug we were having: if new plot options
were chosen (such as a different color map), sometimes when moving
around the previous choices would be drawn.

We moved away from animate with requestAnimationFrame because (on my
laptop at least) it causes the computer fan to come on. So I think
the bug was caused by using old-rendered versions of the scene.
(I'm open to being wrong about this...so we'll have to keep an eye
out for this bug).