Fix render bugs & highlighting point under cursor
Fixed a rendering bug: when the user would change plot options (e.g.
point color) sometimes the scene would revert to the old version as
the user was manipulating the scene. This was caused by not clearing
out the old scene before redrawing.

Also fixed the highlighting: the initial version in this branch was
changing the color of the whole group red (this was just to test things
out, and was never intended as the final version). This commit uses
the transparent sphere to highlight the data point under the cursor.
This information is then sent to the NGCHM to highlight the point
on the heatmap.