Add option for axes at origin (compass rose)
Removed the 'addOriginAxes' because no one liked it. Removed both the
function to create them, and the option from the Gear Menu.

These origin axes here have been essentially replaced by an option for
displaying axes at the origin that act as a compass rose. These are
shown in addition to the box axes, rather than as an alternative.

The origin axes (that have been removed) were drawn as lines. These
new 'compass rose' axes are instead drawn as cylinders and cones
because this gives more flexibility in choosing widths, textures, etc.