Fix selection bug on scene redraw
Added code to explicitly remove scene children in clearScene. Not
having this code was causing a bug where when user would change
plot parameters (such as coordinates) and drag-to-select, spheres
from the old rendering were being selected.

Made changes to the 'demo-plot.html' that helped in resolving the
issue.

Made a small css change.