Save and explicitly initialize OrbitControls
At the end of every OrbitControl event (i.e. rotation/zoom) save
the r, theta, and phi of the coordinate system. In the initialization
of the scene, if an r, theta, and phi are defined, set the OrbitControl
angles to those values. If they are not defined, set them to reasonable
initial values.

This is done to facilitate resetting of the rotation/zoom level when
the user changes things such as the color of the points.

Addresses issue #9.