Add option to choose legend hover transparency
Add an option to the gear menu to allow the user to choose the level
of transparency for the data points corresponding to non-hovered legend
items. Including options for 'Complete' and 'None' (i.e. completely
transparent and not transparent, respectively).

If transparency is turned on, then the non-hovered legend items
themselves are also made transparent (with a fixed reasonable default).
If transparency is turned off, i.e. 'None', then the non-hovered
legend items are NOT made transparent. (because some people dislike
flashing things as they move the mouse around).

Addresses issue #17