Make legend spacing/scrolling improvements
Put legend div and canvas in a flex container and made other CSS
changes in order to improve spacing and legend visibility (legend was
cut off horizontally even if user increased window width).

Added CSS to force a scrollbar in the legend div on *webkit* browsers.
This resolve the problem of: legend is scrollable, but mac user cannot
tell it is scrollable. (So far it seems there's no simple css solution for
Firefox).