Add mouseout listeners for selection box
Added mouseout event listeners for both the scatter plot canvas and the
helper.renderer DOM element. This is to address the situation where the
user is dragging to select points and moves the mouse off the canvas.

Added CSS to prevent user from selecting in icon box and legend-this
was done so that if user moves mouse off canvas while dragging they
do not inadvertently select these items.

Added check that helper box parent element exists before removing
it to prevent console errors.

Addresses issues #22 and #23.