Refactor removing scene elems for performance
Created a new function to remove scene items with the goal of improving
performance as the user drags to select points.

As the user drags to select, we need to show which points they are
selecting during the drag. This requires removing elements from the
scene during the drag. Before this commit, we did this by calling
getObjectByUserDataPropery in a while loop. getObjectByUserDataProperty
is itself recursive. Profiling suggested that whole bit of code was
causing performance issue. So replaced these calls to a cleaner
function for removing scene elements: removeObjectsByUserDataProperty.

The function getObjectByUserDataProperty is kept because it's used
in other places (where it doesn't cause a performance bottleneck).

This may partially address issue #30.