Refactor point selection to improve performance
Change the point selection: previously, selecting points (i.e. by
dragging a selection box) would create a new THREE.Mesh sphere for each
selected point and add it to the scene. Unselecting points would remove
that THREE.Mesh. This commit refactors the selection of points to
instead change the material and geometry of the existing THREE.Mesh
data points to be that specified by the
Plot3D.geometriesMaterials.selection object.