Use sessionStorage to save rotation/zoom state
This project was desiged to be in an iframe in the NGCHM page
(https://github.com/MD-Anderson-Bioinformatics/NG-CHM), as such this
project's web content is sometimes moved around the DOM in the NGCHM
page. This movement causes the page to be reloaded, losing the
current state of the page, including the rotation/zoom level.

In order to keep the same rotation/zoom level upon reload, the spherical
coordinates (i.e. the rotation/zoom level) are saved to sessionStorage,
using the nonce as a key. When one of these reload events happens, a
check of the sessionStorage is performed. If available the
rotation/zoom is set from there. If no values are available, reasonable
defaults are used.

Addresses #32.