Add functions to send/receive plugin data
In the NGCHM project,
https://github.com/MD-Anderson-Bioinformatics/NG-CHM, we are working on
a feature to save and recover a view of the NGCHM. For this, we need the
NGCHM to be able to save all the data required to display a given state
of this plugin. Some of this data is specific to the plugin--i.e. it did
not come from the NGCHM originally.

This commit adds functions to 1) send that data (data required to create
a given view of the scatter plot, but did not come from the NGCHM) to
the NGCHM and 2) use that data if sent by the NGCHM (i.e. when the NGCHM
is reconstructing a saved state).